% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.apply_diff_op}
\alias{.apply_diff_op}
\title{Perform data diffusion}
\usage{
.apply_diff_op(gex, pcs, aff, dft, t_max, tol, exact_solver)
}
\arguments{
\item{gex}{\strong{matrix or Matrix} Gene expression matrix}

\item{pcs}{\strong{matrix} Principal components matrix}

\item{aff}{\strong{dgCMatrix} Markov affinity matrix}

\item{dft}{\strong{NULL or integer(1)} Diffusion time}

\item{t_max}{\strong{integer(1)} Maximum diffusion time}

\item{tol}{\strong{numeric(1)} Tolerance for Procrustes disparity}

\item{exact_solver}{\strong{logical(1)} Perform imputation in gene space}
}
\value{
\strong{list(matrix, integer(1))} Imputed matrix and diffusion time used
}
\description{
Apply diffusion operator to data in order to perform imputation.
}
