% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.calc_diff_op}
\alias{.calc_diff_op}
\title{Compute diffusion operator}
\usage{
.calc_diff_op(pcs, knn, ka, dist_metric)
}
\arguments{
\item{pcs}{\strong{matrix} Principal components matrix (used for kNN search)}

\item{knn}{\strong{integer(1)} Number of nearest neighbors to search for}

\item{ka}{\strong{integer(1)} Number of nearest neighbors to use for adaptive kernel}

\item{dist_metric}{\strong{character(1)} Type of metric to use for distance calculations during kNN search}
}
\value{
\strong{dgCMatrix} Markov affinity matrix
}
\description{
Calculate graph diffusion operator (Markov affinity matrix).
}
