% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.calc_pca}
\alias{.calc_pca}
\title{PCA wrapper}
\usage{
.calc_pca(gex, npc, scale)
}
\arguments{
\item{gex}{\strong{matrix or Matrix} Gene expression matrix}

\item{npc}{\strong{numeric(1)} Number of leading principal components to compute}

\item{scale}{\strong{logical(1)} Whether to scale genes to unit variance}
}
\value{
\strong{list} PC loading/rotation matrices as well as centering/scaling vectors.
}
\description{
Calculate leading principal components (via truncated singular value decomposition).
}
