% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.check_params}
\alias{.check_params}
\title{Parameter validation}
\usage{
.check_params(args)
}
\arguments{
\item{args}{The arguments to the magic_impute function}
}
\value{
NULL (but stops execution for invalid parameters)
}
\description{
Checks that all the parameters used in MAGIC imputation are permissible.
}
