% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.gene_indices}
\alias{.gene_indices}
\title{Finds row indices of signature genes}
\usage{
.gene_indices(mat, gss)
}
\arguments{
\item{mat}{\strong{matrix-like} Gene expression data (genes x cells)}

\item{gss}{\strong{named list of character} Signature genes (with same nomenclature system as \code{mat})}
}
\value{
\strong{integer} 0-based indices (for passing to Rcpp function) of signature genes
}
\description{
For converting from character to integer based indexing
}
