% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.invert_pca}
\alias{.invert_pca}
\title{PCA inversion}
\usage{
.invert_pca(pcs, rot, ctr, sdv, low_mem)
}
\arguments{
\item{pcs}{\strong{matrix} The principal components (scaled left singular vectors)}

\item{rot}{\strong{matrix} The rotation matrix (right singular vectors)}

\item{ctr}{\strong{integer} The centering vector}

\item{sdv}{\strong{integer or NULL} The scaling vector (or NULL if no scaling was applied)}

\item{low_mem}{\strong{logical(1)} Whether to use delayed operations to reduce memory usage}
}
\value{
\strong{matrix or DelayedMatrix} \code{rot \%*\% t(pcs) * sdv + ctr}
}
\description{
Reverses operations done for PCA: back-rotation, unscaling, and uncentering.
}
