% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.minmax_scale}
\alias{.minmax_scale}
\title{Scale vector to unit range = (0, 1)}
\usage{
.minmax_scale(x)
}
\arguments{
\item{x}{numeric: values to be scaled}
}
\value{
numeric: scaled values
}
\description{
Scale vector to unit range = (0, 1)
}
