% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.normalize}
\alias{.normalize}
\title{LogCP10K transform}
\usage{
.normalize(gex)
}
\arguments{
\item{gex}{\strong{matrix or Matrix} Gene expression matrix (cells x genes)}
}
\value{
\strong{matrix or Matrix} Transformed (normalized) matrix
}
\description{
Simple normalization method for scRNA-seq data.
}
