% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.procrustes}
\alias{.procrustes}
\title{Procrustes disparity}
\usage{
.procrustes(x, y)
}
\arguments{
\item{x}{\strong{matrix}}

\item{y}{\strong{matrix}}
}
\value{
\strong{numeric(1)} Procrustes disparity between input matrices.
}
\description{
Calculates symmetric Procrustes distance (adapted from MATLAB procrustes).
}
