% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Performs MAGIC imputation}
\usage{
impute(
  gex,
  transpose = TRUE,
  do_norm = FALSE,
  pca = NULL,
  npc = 100L,
  scale = TRUE,
  knn = 16L,
  ka = 6L,
  dist_metric = "euclidean",
  dft = NULL,
  t_max = 16L,
  tol = 0.001,
  exact_solver = TRUE,
  conserve_memory = FALSE,
  env_ret = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gex}{\strong{matrix or Matrix} Gene expression values (that has passed quality control).}

\item{transpose}{\strong{logical(1)} Whether to transpose gex (make it cells x genes) prior to downstream operations.}

\item{do_norm}{\strong{logical(1)} Whether to perform LogCP10K normalization on gex.}

\item{pca}{\strong{matrix (cells x PCs) or NULL} Precomputed principal component matrix (or NULL to derive it from gex).}

\item{npc}{\strong{integer(1)} Number of principal components (min = 1) to calculate.}

\item{scale}{\strong{logical(1)} Whether to scale columns of input matrix (after centering) to unit variance prior to PCA.}

\item{knn}{\strong{integer(1)} Number of nearest neighbors (min = 2) to consider during distance calculation.}

\item{ka}{\strong{integer(1)} Number of nearest neighbors (min = 2, max <= knn) to use for the adaptive kernel.}

\item{dist_metric}{\strong{character(1)} Type of metric to use for distance calculations during kNN search.}

\item{dft}{\strong{NULL or integer(1)} Automatic (NULL) or user-defined (integer) diffusion time (min = 1, max = 16).}

\item{t_max}{\strong{integer(1)} Maximum diffusion time to test when using automatic diffusion time (min = 1, max = 16).}

\item{tol}{\strong{numeric(1)} Threshold for Procrustes disparity (min = 0, max = 1) between successive diffusion times.}

\item{exact_solver}{\strong{logical(1)} Whether to perform imputation in gene space (TRUE) or PCA space (FALSE).}

\item{conserve_memory}{\strong{logical(1)} Whether to avoid allocating a large dense matrix when \code{exact_solver = FALSE}.}

\item{env_ret}{\strong{logical(1)} Return all variables in the environment (TRUE) or just the imputed matrix (FALSE).}

\item{verbose}{\strong{logical(1)} Whether to issue print statements at different major parts of the algorithm.}
}
\value{
\strong{matrix-like or list}
If \code{env_ret = FALSE}, then just the imputed matrix.
Otherwise the function environment as a list containing all parameters (possibly modified) as well as
\itemize{
\item imp \strong{matrix or DelayedMatrix} Imputed matrix.
\item aff \strong{dgCMatrix} Markov affinity matrix (graph diffusion operator).
\item pca \strong{list} Possibly computed (if pca was NULL), yielding a four element list, where:
\itemize{
\item \code{x} \strong{matrix (cells x PCs)} The principal components matrix (scaled left singular vectors).
\item \code{v} \strong{matrix (genes x PCs)} The rotation matrix (right singular vectors).
\item \code{center} \strong{integer (cells)} The centering vector.
\item \code{scale} \strong{integer (cells) or NULL} The scaling vector (or NULL if no scaling was applied).
}
}
}
\description{
Calculates a graph diffusion operator for the given input matrix and applies it to produce an imputed matrix.
}
