% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wks}
\alias{wks}
\title{Helper function for performing a weighted Kolmogorov-Smirnov like procedure.}
\usage{
wks(gene_ranks, sig_idx, aux)
}
\arguments{
\item{gene_ranks}{numeric: centered gene expression ranks}

\item{sig_idx}{numeric: indices of genes belonging to the signature (gene set) being scored}

\item{aux}{numeric: two element vector containing
0. Centering value that was used for ranks
\enumerate{
\item Start index (after centering) for the running sum procedure
}}
}
\value{
Modified Kuiper statistic (sum of minimal and maximal deviations during running sum)
}
\description{
Helper function for performing a weighted Kolmogorov-Smirnov like procedure.
}
