% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.gene_indices}
\alias{.gene_indices}
\title{Finds row indices of signature genes in given gene expression matrix}
\usage{
.gene_indices(mat, gss)
}
\arguments{
\item{mat}{matrix-like: gene expression data (genes x cells)}

\item{gss}{named list of character: signature genes (with same nomenclature system as mat)}
}
\value{
0-based indices (for passing to Rcpp function) of signature genes
}
\description{
Finds row indices of signature genes in given gene expression matrix
}
