% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.minmax_scale}
\alias{.minmax_scale}
\title{Minmax transform}
\usage{
.minmax_scale(x)
}
\arguments{
\item{x}{\strong{numeric} Values to be scaled}
}
\value{
\strong{numeric} Minmax transformed values
}
\description{
Scales input vector to unit range
}
