% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Seqtometry scoring}
\usage{
score(mat, signatures, minmax = TRUE)
}
\arguments{
\item{mat}{\strong{matrix-like} Gene expression data (genes x cells)}

\item{signatures}{\strong{named list of character} Signature genes (with same nomenclature system used in \code{mat})}

\item{minmax}{\strong{logical(1)} Whether to perform minmax transform on scoring results (default: TRUE)}
}
\value{
data.table (cells x signatures): single cell scores for each signature, where
cell barcodes are stored in the "id" column
}
\description{
Scores signatures for the given matrix (which has already undergone QC, normalization, and imputation)
}
