\name{Seqtometry-package}
\alias{Seqtometry-package}
\alias{Seqtometry}
\docType{package}
\title{
  Signature scoring for single cell analysis
}
\description{
  This package provides the Seqtometry scoring algorithm for the generation of single cell 
  gene set (signature) enrichment scores for scRNA-seq or scATAC-seq data.
}
\details{
  Call the score function on a gene expression matrix that has undergone quality control,
  normalization, and imputation using the desired collection of signatures.
}
\author{
Robert Kousnetsov <robert.kousnetsov@health.slu.edu>

Maintainer: Robert Kousnetsov <robert.kousnetsov@health.slu.edu>
}
