\name{Seqtometry-package}
\alias{Seqtometry-package}
\alias{Seqtometry}
\docType{package}
\title{
  Signature scoring for single cell analysis
}
\description{
  This package provides functions used in Seqtometry (Kousnetsov et al. 2024), a method for analyzing
  single cell (scRNA-seq or scATAC-seq) data via signature (gene set) enrichment scores.
}
\details{
  Call the score function on a gene expression matrix that has undergone quality control,
  normalization, and imputation using the desired collection of signatures. Visualize the
  signature scores and repeat the process as needed to gain insight into the data.
}
\author{
Robert Kousnetsov <robert.kousnetsov@health.slu.edu>

Maintainer: Robert Kousnetsov <robert.kousnetsov@health.slu.edu>
}
