% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wks}
\alias{wks}
\title{Helper function for performing a weighted Kolmogorov-Smirnov-like procedure.}
\usage{
wks(gex, gss, mus, sds)
}
\arguments{
\item{gex}{numeric: normalized gene expression values}

\item{gss}{list: indices of genes in each gene set}

\item{mus}{numeric: means of all genes}

\item{sds}{numeric: standard deviations of all genes}
}
\value{
Modified Kuiper statistic (sum of minimal and maximal deviations during running sum)
}
\description{
Helper function for performing a weighted Kolmogorov-Smirnov-like procedure.
}
